Improve deduplication of issues for Cx
It's possible that the issue with the same name will go to different query results
in Cx report. As a result, we display several identical issues in one jira
ticket (group and category will be different though).
Deduplication key is updated to contain the name of the issue.
Test data is updated correspondingly.